requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.85
botocore==1.29.85
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==4.1.7
django-allauth==0.44.0
django-ckeditor==6.5.1
django-cors-headers==3.6.0
django-countries==7.5.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-storages==1.13.2
echo==0.8.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==2.10
jmespath==1.0.1
jsonfield==3.1.0
numpy==1.24.2
oauthlib==3.1.0
packaging==23.0
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
QtPy==2.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
s3transfer==0.6.0
six==1.15.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
stripe==5.2.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.2
whitenoise==5.2.0
